Add files via upload
A script that I threw together that will batch download files that are not directly accessible through NZQA's website (though are still on their web server)- mainly marking schedules from 2017-2019. The links are easy to guess, (e.g change the occurrences of 2020 to 2019) but this just saves you from having to repeatedly change the link in your browser. It barely works- you may have to adapt it to make it function but feel free to use it however you wish.